Update patch set 2
Patch Set 2: Code-Review+1

The failing job is verifying that a all jobs triggerable by Zuul are actually deployed in Jenkins.  Since the jobs do not exist yet (they need to be deployed from this change), the layout validation job fails.

The behavior is changed with: https://gerrit.wikimedia.org/r/#/c/170320/ which would no more verify whether jobs are deployed on Jenkins on patch submission but would on +2.  That will be less confusing.


Change is good to go.

Patch-set: 2
Label: Code-Review=+1