Update patch set 3
Abandoned

The change as proposed is not going to work since Zero is a production portal with authentication, which is not what is expected (all other wikis are public).

A solution has to be figured out by mobile to be able to run browser tests against the production wiki from the Jenkins host.  Follow up on https://bugzilla.wikimedia.org/show_bug.cgi?id=72326 (and see my comment #2 there).

Patch-set: 3
Status: abandoned