Update patch set 6
Patch Set 6:

The job triggered by the postmerge pipeline above had the parameters:

 DOC_SUBPATH=master
 BRANCH=master

I have hacked the Zuul layout configuration to trigger the publish job via the 'publish:' pipeline which reacts on tag and pushed tag 1.0. The job had the parameters:

 DOC_SUBPATH=1.0
 ZUUL_REF=ref/tags/1.0


So it is all good.

We can now use DOC_SUBPATH in the publish jobs to namespace the builds.

Patch-set: 6
Label: Verified=0