Add customization of publishers for composer jobs
A composer command might generate files which we want Jenkins to
interpreate via a publisher.  The cdb project runs phpunit which creates
a XML report of the tests.  We want to be able to use the xunit
publisher for this command but not for others (since they do not
generate the xunit report).

Add to the job template `{name}-composer-{composer_command}` a parameter
`composer_publishers` which is a sequence of publishers to add to the
job templates.  We are using a !!python/tuple to prevent JJB from
realizing a job per element of othe sequence and pass them as is to the
job template.
http://ci.openstack.org/jenkins-job-builder/definition.html#variable-references

Add a default: composer_publishers: !!python/tuple []

Make the cdb composer phpunit job to use the publishers 'phpunit-junit'
and 'archive-log-dir'.

We probably want to use the same trick for the python tox and ruby
bundle jobs.

Change-Id: I632440782c5b838243f7d0fcf8b424d5a43d8d12