Publish doxygen doc for the 'cdb' project
First use of the '{name}-doxygen-publish' job template introduced
previously. Should make the Doxygen documentation published under
https://doc.wikimedia.org/cdb/<VERSION>/.

In Zuul, that is the second repo ever to have a job in the 'publish'
pipeline which reacts on tags. Not sure whether the 'push-doc' job
handles ref-updates properly but we will now be able to test it.

Bug: 73530
Change-Id: I591750a7defa13625ea0f440394ca34ab622913c