Update patch set 1
Patch Set 1:

(1 comment)

The use of `composer install` will by default install all packages in require-dev. For many projects this will include phpunit of some variety which we probably don't need. For the cdb project in particular there is a very loose `"phpunit/phpunit": "*"` dependency in require-dev that is going to pick up each and every new phpunit release and add it to the shared composer package cache.

I'd really almost lean towards standardizing on a single "composer jenkins" script command to be managed in the package or something similar to .travis.yml that we can add to manage the needed steps on a package by package basis.

Patch-set: 1
Label: Code-Review=0