jsduck: Remove redundant hack that searched jsduck.log for "Warning"
Follows-up a4db3e7a37, e7a940d94b.

When we added --processes=0, this hack became redundant since the
upstream issue with exit codes only applies to parallel execution
on old Ruby versions.

As long as we run the old Ruby (bug 60138) we'll have processes=0,
and when we upgrade, the exitcode bug isn't there and we'll remove
processes=0.

Bug: 55668
Change-Id: I1e226492d5313e28d851d71d04771fd84bd66af8