Update patch set 1
Patch Set 1:

When a job using npm causes a failure, there should already be a sufficient level of verbosity in the console output. npm-debug.log is mostly for later reference. If you found it useful it the past, it's fine by me, but just checking whether it's theoretical or not.

Patch-set: 1
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0