Update patch set 1
Patch Set 1:

(1 comment)

We could use 'test' per default and pass the parameters in an env variable that will only be applied on the PHPUnit command.  See inline comments for a proposal.

If that works, I will rework the serie of JJB patches and update the change pending for the cdb project (the last patch uses 'phpunit' as an entry point).

Patch-set: 1