Soften branch filter for publish jobs
To be able to reuse the exact same job in postmerge and publish
pipeline, we need tweak the branch filter which is usually set to
'^master$', the reason is that the event for a new tag lack a branch and
Zuul fallback to matching the ref (ex: refs/tags/1.24.0).

Lot of publish jobs have branch: '^master$', that can be removed as we
add them to their projects publish: pipeline.

T75991

Change-Id: I3f06f8f9101a29949741171dfdce0d7585c90dc5