Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

For Zuul, you will want to use to use templates that will generate the proper pipelines/job names.  That will add the test extension job which runs mediawiki core PHPUnit tests as well as the extension tests (if there is any).

You also need to have jobs generated in Jenkins, that can be done by editing jjb/mediawiki-extensions.yaml , find the huge list of extensions and insert this extension name in it.  That will realize all the jobs needed.


Looking at the code, it seems that provides a client side JavaScript lib to interact with the Wikibase API provided by MediaWiki.  I am not sure why it is an independent repo. Not a problem though :]

Patch-set: 1
Label: Code-Review=-1