Update patch set 2
Patch Set 2:

I will check and remove those that are not used anymore in a new patch.

I'm not in the wmde LDAP group (nobody thought of adding me). I also prefer not to use such a group as it is organization centric. But I am in the nda LDAP group (used by ops for e.g. logstash, graphite, etc. access). Could you give the necessary permission to my account or if you prefer to the nda group?

How is provisioning of new slaves broken? Is there some ticket for that?

Patch-set: 2
Label: Verified=0