Update patch set 1
Patch Set 1: Code-Review-1

> These jobs should only run under python3...do I need to specify that somewhere?

tox comes with default environment (py26, py27, py33, py34) that are bound to a specific python version.  In the tox.ini file you can rename [testenv:channels]  to [testenv:py34]  and adjust the env name in the JJB/Zuul files.  The invocation will then be:

 tox -epy34

Tox will look for `python3.4` and set the env with it. On our Jenkins slaves, that require the job to run on Trusty (which has python3.4) which is what '{name}-tox-{toxenv}-trusty' does :]

If you want to preserve 'channels', you can add the python version to it:

 [testenv:channels]
 basepython = python3.4

Just using the convention of 'py34' is probably better.


>  The channels test depends on the PyYAML package, does anything special need to be done to have that installed or will tox take care of it?

You need to define the dependencies that pip is going to install, looks like you did already in the repo:

 [testenv:channels]
 commands = ./test_channels_yaml.py
 deps = PyYAML

So looks fine.


There is some documentation at: https://www.mediawiki.org/wiki/Continuous_integration/Tutorials/Test_your_python Feel free to amend it.



-1 cause the 'channels' env needs either to be renamed to 'py34' or use 'base python = python3.4'.   Otherwise good.

Patch-set: 1
Label: Code-Review=-1