Update patch set 1
Patch Set 1:

[integration-zuul-layoutvalidation-gate] 04:29:12 Job citoid-npm not defined

Zuul can validates the jobs defined in the layout file against an arbitrary list of jobs.  The -gate job retrieve the list of Jenkins jobs and bails out when one is defined in Zuul but has not been deployed in Jenkins yet.

That is to prevent people from merging/deploy an invalid layout file.

Patch-set: 1