Update patch set 1
Patch Set 1: Code-Review-1

If they're for "dev scripts" only (whatever that is), then the problem is in MobileFrontend. By default npm-test does nothing. Whatever it does is configured in MobileFrontend package.json and Gruntfile.js. It should do linting of javascript, which is mandatory. If that is not passing, that should block a commit without question.

If there's to many unrelated tasks being executed, update package.json and/or Gruntfile so move them out of the 'test' alias and into a separate task developer can run locally when they choose. Thus ensuring that 'test stays on-topic and useful.

Patch-set: 1