Update patch set 4
Patch Set 4: Code-Review-1

Patchset 3 removed the use of Zuul cloner when I want to eventually get rid of the Jenkins git plugin entirely.  I suspect Timo did it because the Jenkins plugin clean the workspace unlike Zuul, but composer would clean the packages that are no more in the composer.json so it is probably not needed.

https://gerrit.wikimedia.org/r/#/c/174410/2..4/jjb/job-templates.yaml

Adding in the 'composer validate' is a good idea :)


We still have the issue of not being able to generate the junit.xml file though :-\

Patch-set: 4