Update patch set 3
Patch Set 3:

From my comment on the task:

The Jenkins job templates are all expecting composer.json to be a the root of the repository. Can we move it there to avoid adding exceptions to the convention? We could use it to install PHPUnit and run tests as well and I would really prefer to avoid having multiple composer.json files per repo.

So I guess we will want to move the composer.json to the root of the operations/mediawiki-config.git repo or we would need a custom job :( .

Patch-set: 3
Label: Verified=0