Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

The web service is light weight. I like it :-] See inline diff for a few required changes.


We can get the job to only run from gallium.wikimedia.org which has the IP 208.80.154.135 . In the job-template:

 node: gallium

Then ensure in your script that the source is 208.80.154.135 .

If you go this way, please add a comment that the remote web service whitelist that IP. This way when we phase out gallium we will remember to get the whitelist adjusted.

Patch-set: 2
Label: Code-Review=-1