Update patch set 3
Patch Set 3:

> I just realized we cannot IP-whitelist, because IPs are not available in tool labs.

Sorry I should have thought about it yesterday. The host tools.wmflabs.org is actually a proxy, so your instance see the IP of the proxy and not the original request.

The generic labs proxy (but most probably the tools labs web proxy as well), should set a X-Forwarded-For header which contains the IP of the original client. Example:

 curl http://integration.wmflabs.org/?date=`date +%s`

Generates on the instance the request:

 Host: integration.wmflabs.org
 X-Forwarded-Proto: http
 Connection: close
 X-Forwarded-For: <MY IP ADDRESS HERE>
 User-Agent: curl/7.30.0
 Accept: */*

Patch-set: 3