Support injecting an extension in shared job
The mediawiki-extensions-{hhvm,php} jobs have an hardcoded list of
extensions to be tested together. When we want to add a new extension,
we want to be able to try the job with it before including it.

Add mediawiki-extensions-hhvm to extensions 'experimental' pipeline.
Adjust the job to inject the project of triggered patch in the list of
dependencies to be tested.

Change-Id: I7467f71910ac5fae7055ea675d6af20f9ccc8e4d