mediawiki-core-jsduck-publish: Add trailing slash in rsync commmand
Follows-up c406e87. These were needed as otherwise rsync will
copy resources to a subdirectory of modules instead of to modules
itself (resources/* -> modules/*, instead of resources/modules/*).

For the 'rm' command it has to be without trailing slash however
as otherwise it fails with "rm: cannot remove: Is a directory"

Bug: T86864
Change-Id: Idad04d85d6423ff3dc02ebaeb55c8cf01dbaac85