mediawiki-core-code-coverage: Fix cloverphp reports in Jenkins
Since the job was converted to use Zuul-cloner, the clover
log ended up in /src/log instead of /log.

Fix the phpunit command to send the clover.xml file to /log instead
of /src/log so that it can be picked up by Jenkins.

The html report functioned properly throughout as that one
is exported directly from phpunit to /srv/org/wikimedia/integration/cover

Change-Id: I1e615b7fa4b1b6f5331b0774dc3984517bd3fb22