Update mediawiki-core-qunit to use mw-set-env-qunit.sh
For the old slave-scripts/wmfgrunt-based job these aren't strictly
needed as it can know the localhost already (though it's better to
not duplicate this between jobs).

This in in preparation for the Karma-based job where all logic
is inside MediaWiki core, which shouldn't have knowledge of the
Jenkins-specific install as it would differ from developer's
local install.

Change-Id: Ib722b79777e700f8142e9b5b2d468bccc097630e