Ensure extensions are loaded in a consistent order
When using the experimental pipeline, the triggering project is appended
at the end of the list extensions to load.  While testing the AntiSpoof
extension, the tests passed just fine, but when adding it to the list of
extensions they will start failling.

To ensure experimental behave the same as the test job, sort the list of
dependencies to load ensuring both are consistent.  The AntiSpoof
experimental job now fails as expected.

Updates:
    mediawiki-extensions-hhvm
    mediawiki-extensions-zend

Change-Id: I94f65b5636871bf92202115bed3f11226ccb9448