Update patch set 2
Patch Set 2:

When using 'check experimental' on the AntiSpoof extension, it ends up being loaded last and that has a side effect with some other extensions. This patch merely make check experimental consistent with the order we will end up with when adding the project to the shared job.

I replied on T88015 about ZeroPortal die() and proposed a patch (tip: use SetupAfterCache hook).

Patch-set: 2