Migrate cxserver back to regular npm job
In eb2a61bef I recommended Kartik to reuse for cxserver the same logic
we had for the parsoid service.  Turns out it is overkill since cxserver
only uses 'npm test' and all the cruft for parsoid is due to them having
multiple tests.

Move cxserver back to the generic '{name}-npm' template which just
invokes 'npm test'.  That is nicer.

In a next change, I will create a job/template for the deploy
repository.

Change-Id: I2bdb7f34f554476a7e1910f4de63452cfaba3a89