Update patch set 1
Patch Set 1:

(1 comment)

Hey Matanya, good to see you around :]

The bug header in the commit message is not triggering the comment spam in Phabricator, it needs to be above Change-Id see inline diff.

The change is on hold till ops team wants to actually have it voting (I have poked them on the ops list).

Patch-set: 1
Label: Code-Review=0