Update patch set 3
Patch Set 3:

Good catch Chris, I forgot about the EAL :-(

We still export three variables:

 MEDIAWIKI_USER= ... whatever user name
 MEDIAWIKI_PASSWORD= ... whatever password
 MEDIAWIKI_PASSWORD_VARIABLE='MEDIAWIKI_PASSWORD'

https://gerrit.wikimedia.org/r/#/c/189938/3/jjb/macro-browsertests.yaml

It is just that the password are held in the Credentials store provided by a plugin instead of in the global configuration and exposed to all jobs.

With EAL, I guess we will have a shared environments.yml file and tweak it to inject the passwords we need.  The Credentials plugin apparently can hold such a file already populated with the passwords.

Patch-set: 3
Label: Verified=0