Update patch set 3
Patch Set 3: Code-Review-1

That fix the dependency issue for Echo.  Now ParserFunctions could receive a change which pass its tests but could have some side effects on the other extensions.  To prevent that, we need to trigger the shared job for patches made against ParserFunctions.

Additionally, the tests most probably do not work on the old release branches. So we do not want to use the shared job.

An example for the Translate extension is:
https://gerrit.wikimedia.org/r/#/c/188518/1/zuul/layout.yaml,unified

Just add the 'extension-gate' template to ParserFunctions. Then filter out the 'standalone' job for all release branches, it is no more needed in master branch since the extension will trigger the shared job.

Patch-set: 3
Label: Code-Review=-1