Pass MEDIAWIKI_CAPTCHA_BYPASS_PASSWORD to GettingStarted browser test
Add to the browsertests default a captcha password used by the
GettingStarted browser test job.

We could have used a more specific job template, but defining a wrappers
section in it would override the wrappers defined in the default. Copy
pasting the default is asking for maintenance troubles down the road, it
is much easier to populate the variable to all jobs at the default
level.

Bug: T91220
Change-Id: I8bbff7c6b95b33c2dad81905f876c464d118cb6c