Update patch set 1
Patch Set 1:

For the main 'wikimedia-fundraising-crm' job this would ignore the submodules and use latest master. Which will most likely lead to two issues:

1) Making changes in one of the subrepos, that requires a change in the parent repo can't be tested/merged since there's an unnatural downstream/upstream enforcement being made.

2) If the change is not dependent but you forget to update the submodule, it will work in testing but break in production.

Depending on the kind of project this is and how it manages submodules, this may not be an issue, but I've seen it go wrong in other projects so you may not wanna go down this route of using zuul-cloner to fetch the submodules (instead of using 'git submodule update --init'. And may want to leave out testing with upstream in the downstream repo. It means you'll catch regressions a few minutes later, but you'll still catch them (in the parent repo when drafting the commit to update the submodule, it will test it still). And also keeps a more atomic and standalone integrity.

The only exception to this if you never make breaking changes that require consuming software to make changes in the same release (e.g. always maintain 1 window of back-compat so that you can add the new in the downstream repo, update the parent repo, update the callers, remove old from downstream repo, update parent repo etc.).

Patch-set: 1
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0