Update patch set 1
Patch Set 1:

(3 comments)

At first I thought we should use scap, but it has proven to be annoying.  In this case the fabric part is only running on our local machines so it does not really introduce a new deployment system, it is merely another way to write a list of commands.

zuul.eqiad.wmnet is a nice trick, but I would rather point to the canonical hostname gallium.wikimedia.org.


pip install --user fabric  :-D


No need to change anything, that is good to me thanks. Will let other reviewers look at it as well.

Patch-set: 1
Label: Code-Review=0