Don't use separate ssh-agent for running scap
Since https://gerrit.wikimedia.org/r/#/c/195340/ landed,
beta also uses keyholder (just like production) to do deploys.
Jenkins seems to be using the jenkins-deploy user to be running
the scaps, and jenkins-deploy user is a member of the
project-deployment-prep group, which is authorized to run
scap via keyholder. So the wrapper script is unnecessary now
and actually causes failures.

Change-Id: Ia215d94fd4d15aef097d47697e4fbaba20391557