Update patch set 2
Patch Set 2:

Krinkle Fabric has a few dependencies, one of them is pycrypto which is a C extension to python (a compiled module). Most of the output is from the configure and gcc One could have gotten rid of it by installing the packages available in the distribution (on Debian ptyhon-paramiko should bring everything).

Patch-set: 2