Update patch set 4
Patch Set 4:

And it's more about the reload than the deployment of a change. It communicates to team members that the queue will be stalled for a while and that one probably shouldn't do stuff like restarting Jenkins, Gearman or Zuul at the same time.

Also, merges in zuul/ are not logged in SAL either. So that's not helpful to others and would involve opening gitlog all the time.

Patch-set: 4
Label: Verified=0