Update patch set 1
Patch Set 1:

In case the Zend test is failing on +2, a new patchset would be sent and I guess +2 ed to have it merged.

The change is fine and I like the idea, but I am not of the usefulness of being able to trigger the Zend tests on demand.

Patch-set: 1