Stop throttling SauceLabs jobs
The browser tests once executed directly on the slaves and we suspected
some performance bottleneck when several ran concurrently.

We found out there is a race condition that causes the XVFB server to be
killed and causing the failure of other jobs running on the same job

Thus, we wanted to only have one browser test per job.

Nowadays we have migrated all browsertests to run on SauceLabs. The part
of code running on the slaves is quite lightweight.  Thus remove the
throttling from all browser tests jobs.

Change-Id: I8f9d34ce19a0b5a878e40dbb79f55f10fd43a7ae