Update patch set 2
Patch Set 2:

Timo is right we might reach out SauceLabs quota.

Zeljkof said there is no point in running the jobs twice per day, so we are going to change the timer to @daily which will spread the browser test runs over the course of 24 hours.

Given the # of jobs we have and our available quota, we are not going to exhaust it.

Patch-set: 2