Update patch set 2
Patch Set 2: Code-Review-1

The 4 hours timeout is merely for the Wikidata job. I am not a fan of having the same timeout for all jobs.  Instead we can introduce a new variable {browsertest-timeout}, default it to 120  and then override it on a per job basis.

I will do the grunt work.

Patch-set: 2
Label: Code-Review=-1