Support per browsertest job timeout
Our browser test jobs all share the same 180 minutes timeout. In most
cases that is way too long and for the WikidataTests job it is too
short.  This patch bring support to specify the job timeout on a per job
basis.

Add a variable 'browsertest_job_timeout' = 180 to the browsertest
default.
Change the timeout wrapper to use that variable. Will let us override it
on a per job basis using something like:

jobs:
 - myjob:
   browsertest_job_timeout: 60  # override to 60 minutes

Change-Id: I2ede6fc8e171a65db71728c64ee17837e29e843e