Make browser tests version '' by default
SauceLabs supposedly grab the latest version of the browser when no
version is specified. This let us clean a bit our huge shell script.

* Remove switch/case that explicitly set the version per browser.
* Set JJB 'version' variable to default to '' instead of 10, seems it
* was used for IE and every jobs already have a version set explicitly.

This will let us pin a specific version of a browser on a per job basis.
For example Chrome @ v 33 if a repository does not work on later
versions.

Bug: T90678
Change-Id: Id6a8bb62a6884c61862b2890cdf86693e3b56248