Update patch set 4
Patch Set 4:

(4 comments)

Thank you Matthew to have crafted the patch. Using a more specific job template is a good idea, unfortunately the wrappers: section it contains will dismiss entirely whatever is defined in the default.  We could copy paste but I am afraid it is going to be a maintenance burden.

Instead I will send a follow up which adds the MEDIAWIKI_CAPTCHA_BYPASS_PASSWORD to the default and thus populate it to all browser test jobs.  That is not ideal, but easier to maintain.

Patch-set: 4
Label: Verified=0