Prevent rubocop on legacy mw/core branches
The mediawiki-core-bundle-rubocop job assume to have /Gemfile at the
root of the repository with rubocop as a gem.  That is only the case in
master and will be in REL1_25. Moreover, rubocop is not in REL1_24
Gemfile.

Update the job branch filter to prevent it from running on legacy
branches.

Bug: T94549
Change-Id: Ia4c62089ad3ad27f3a258f36554792bda09c34be