Update patch set 2
Patch Set 2: Code-Review+1

After thinking about this it is the right solution for this repository. Additionally having the workspace in a defined state would be nice, but we will hopefully get that with the work on isolation.

Is deployed to Jenkins. Please merge.

Patch-set: 2
Reviewer: JanZerebecki <640@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1