Update patch set 3
Patch Set 3:

We have a private list, e2 (this might be too spammy for the public list).  e2 does go to all team members, not just devs, but they might be interested or could filter it out.

I don't have a strong preference, as long as we get this merged.

Patch-set: 3