Stop running Zend tests
Our unit-testing infrastructure is, by common consensus, all three of these
things:

1. Extremely useful for developers
2. Vital for site reliability
3. Slow as shit

We can improve (3) considerably without impacting (2) at all and only minimally
impacting (1). We can and should offload the running of Zend tests to Travis
CI. (We can configure it to report failures more prominently via IRC / e-mail, too).
This is exactly what we did for HHVM tests when HHVM was not the default
interpreter.

Change-Id: Iab5abe79ae2690d15851c5990fc32856c28e5032