Update patch set 1
Patch Set 1:

The MediaWiki core tests being crazy slow has a few recent causes:

* job has been moved from real hardware to labs instance
* the database backend moved from SQLite over tmpfs to MySQL

And of course the usual issue that the mediawiki core tests are badly written and most ends up invoking the whole stack instead of unit testing with mocks.


What we did is get rid of zend for patchsets being proposed. Kunal proposed the 'check zend' command to trigger them manually.

The Zend tests are run in the gate since we do not want a faulty change to land in. That would cause subsequent job runs to fail due to that change.



We can't use any async jobs like we did to try out HHVM on Travis or by using postmerge.  Experience has shown that such jobs failures are barely noticed and/or take a while to get fixed, in turn the job is always in a fail state and people mostly ignore them.   That is more or less what is happening with the browser tests :\

Patch-set: 1