Update patch set 1
Patch Set 1: Code-Review-2

I want to keep the DependentPipeline for a wild range of reasons I mentioned on T94322.

The reason it is broken right now is because the jobs have been made generic names and Zuul as the convention that it means the attached project are tightly coupled together and hence should share the same queue.

There are two ways to fix it:

* change the job names to trick in Zuul in splitting the queues in the DependentPipeline. We could have mediawiki jobs be prefixed with 'mw-' and that would let us with two queues (one for mediawiki related jobs, an other for the rest of jobs)

* hack Zuul source code to stop automatically merging projects whenever they share a common job and instead add a feature to let us explicitly define the queues.


Disabling the DependentPipeline is a poor solution.

Patch-set: 1
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-2