Split up misc *-publish from *-coverage
When publish is triggered for release tags, it should not trigger
coverage since we don't want coverage reports for branches/tags.
And more importantly, right now the destination on the web server
is per-project, not per-version, so it would overwrite the coverage
report with that of a random branch or tag.

Change-Id: Ie309ee0bed8c5f3a9d51ba541b43b4cc3e87ac95