Move integration-zuul-layout* to labs
The integration-zuul-layout* rely on zuul-server being available. Since
Zuul is now shipped by a Debian package (T48552) we can migrate the jobs
to labs instances.

Since the Debian packages slightly vary between distributions, tie the
jobs to Precise which is the distribution used on gallium.

Inline a fake zuul.conf that has an empty [zuul] section.

Bug: T96463
Change-Id: I5bfdc033dcfda70fd3a0d04ce95a0a626f2265a4