Switch Android mobile app from maven to gradlew
Based on instructions by Bernd Sitzmann, build the Android app using
gradle targets 'checkstyle' and then 'clean assembleAlphaDebug'.

Gradle uses a local cache in the user home dir. On labs we have it
mounted to the instances extended disk so there is plenty of place to
download the whole internet.

A build of master branch passed.

* Capture build artifacts (.apk, checkstyle.xml).
* Made the checkstyle publisher to fail the build whenever there is an
  error no matter its priority (low, normal, high).
* Split lintAlphaDebug goal to its own non voting job since it is not
  passing yet.

Creates:
- apps-android-wikipedia-gradlew
- apps-android-wikipedia-gradlew-lint

Bug: T88494
Change-Id: I159ff994acf0f88e3e6a77a8c65c32137a0d6874