Build Android mobile app with gradlew
Based on instructions by Bernd Sitzmann, build the Android app using
gradle targets 'checkstyle' and then 'clean assembleAlphaDebug'.

Gradle uses a local cache in the user home dir. On labs we have it
mounted to the instances extended disk so there is plenty of place to
download the whole internet.

A build of master branch passed.  The job can later be improved to
capture artifacts and attach them to the build result.

Creates:
- apps-android-wikipedia-gradlew

Bug: T88494
Change-Id: I159ff994acf0f88e3e6a77a8c65c32137a0d6874