Fix fundraising deployment branch filters
The DonationInterface and CiviCRM jobs should not be run when a change
is made against the deployment branch.  The regex would never match
anything though: ^(?!deployment)$  which would be ^$  ie does not match
'master'.

Move the beginning and end of string markers inside the parentheses.

Follow-up 08e8076 and dba69f9.

Bug: T94586
Change-Id: Ie2b9020b9ebd1999caa9773bcf5d35357a906c10