zuul: point Android app lint result to HTML report
The apps-android-wikipedia-gradlew-lint uses the Android SDK linter.
It generates a nice HTML report which is quite handy to point humans to.
Update the job URL pattern.

Culpirt: the report is an artifact. If the build fails before the report
is generated, the URL will point to a non existent page.

Bug: T99112
Change-Id: I28b126f341909daead8628fe12f86beb400c87cc