Update patch set 2
Patch Set 2:

Not sure if I've done this right. What I want is for phplint and phpcs to run as `check`.

legoktm told me that the phpcs job is the old way to do it and I should use a composer.json for that instead.

If there are any minor issues with the commit, please correct them.

Patch-set: 2
Reviewer: Polybuildr <2192@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0