Update patch set 1
Patch Set 1:

Amazing! Holding change till Yaron has the opportunity to +1 it :-)

There are a lot more Semantic* extensions that we might want to fix as well. Some might be passing now that we have switched from Sqlite to MySQL as a database backend.

A potential future would be to have a single job triggered by all the Semantic repositories that will test them all together.

Patch-set: 1