Update patch set 1
Patch Set 1: Code-Review+1

Giving a +1 to this change. Thanks again! I'm not really sure what this change will provide, :) but it seems like a step in the right direction.

Hashar - it should be noted that Semantic Forms is distinct from the other "Semantic" extensions in that it doesn't require the presence of Semantic MediaWiki. It used to, which is why it has that name, but at this point it's an independent extension that should be considered separately from the rest.

Patch-set: 1
Reviewer: Yaron Koren <195@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1