Update patch set 3
Patch Set 3: Code-Review-1

As outlined in the past with Antoine, avoid using 'make build' for this purpose. It incurs additional complexities that aren't worth the benefit.

In case of MobileFrontend, it uses a custom setup with options specified in Gruntfile instead of jsduck.json. If those options are moved to jsduck.json and the redundant file-matching removed, it can simply use the 'jsduck' and 'publish' macro as other projects are using already.

Patch-set: 3