Zend Wikibase jobs only triggered in gate-and-submit
On mediawiki/core we no more trigger the Zend jobs when a patch is
proposed. They are only run on CR+2 (gate-and-submit). It is still
possible to trigger the Zend jobs though by commenting check zend on an
open patchset. See https://gerrit.wikimedia.org/r/#/c/197951/ 7e8a37f1

This apply the same mecanism to mediawiki/extensions/Wikibase to save on
processing. The Zend jobs are still run on CR+2.

Bug: T101382
Change-Id: I9e802565052a3b443ba90566a28993ef5bc6578c