JJB: move single use macro in the job-template
We have a few macros in jjb/macro.yaml that are used by a single
job-template. Move the builder directly to the job template. Makes it
easier to find what the job-template is actually doing.

puppet-lint* builders needed curly braces to be escaped since they are
passed to python str.format().

No jobs changed.

Change-Id: I87a0cf9b21fb053903371ecc1060d52ba9983275