Factor VE submodule update in a macro
zuul-cloner does not process submodules but the MediaWiki extension
needs it.

Create a new macro ve-submodules-update to reuse the same code

All extension jobs now have the additional commands:
    git submodule status
    git submodule foreach git clean -xdff -q

The VE jobs are slightly changed. The submodule update only triggers
when mediawiki/extensions/VisualEditor is listed in deps.txt which is
always the case.

Change-Id: I8881c1c03a298f24ff15b452fb2639eddb2d2e58