Create dedicated jslint/phplint instances for apps/* repos
Most commits don't trigger any tests in mobile right now. But are
blocked from merging regardless for a long time due to Zuul
behaviour described in T94322.

Create separate instances for now so that the apps teams are not
motivated to bypass Jenkins, which in turn causes even worse delays
due to Zuul not seeing the merge and holding everything up for
10 minutes or more.

Note: {name}-jslint runs both jshint and jsonlint, so no separate
      jsonlint job needed.

Change-Id: I70dceeef25762a354d2ef730bb0c5fb09f18d03f