Update patch set 1
Patch Set 1:

We can potentially make it a single job that is triggered for all extensions we have on Wikimedia.  Such an effort started with the mediawiki-testextensions-* jobs.

We might even inject the selenium run into the mediawiki-testextensions-qunit job since the setup is very similar.This way the job would:

* setup the env
* run qunit
* bundle install
* cucumber

Thus if qunit fails, we skip cucumber :-}

Patch-set: 1
Label: Code-Review=0