Update patch set 10
Patch Set 10:

I get this error now 

        => Fixing file: 0/1 violations remaining [made 2 passes]... ←[32mDONE←[0
m in 21ms
'diff' is not recognized as an internal or external command,
operable program or batch file.
Processing MetrolookTemplate.php [PHP => 5844 tokens in 733 lines]... DONE in 27
8ms (35 fixable violations)




        => Fixing file: 0/35 violations remaining [made 4 passes]... ←[32mDONE←[
0m in 1.3 secs
'diff' is not recognized as an internal or external command,
operable program or batch file.
Processing SkinMetrolook.php [PHP => 543 tokens in 84 lines]... DONE in 26ms (2
fixable violations)



        => Fixing file: 0/2 violations remaining [made 3 passes]... ←[32mDONE←[0
m in 101ms
'diff' is not recognized as an internal or external command,
operable program or batch file.
Processing SkinMetrolookHooks.php [PHP => 1296 tokens in 191 lines]... DONE in 6
9ms (20 fixable violations)



        => Fixing file: 0/20 violations remaining [made 3 passes]... ←[32mDONE←[
0m in 203ms
'diff' is not recognized as an internal or external command,
operable program or batch file.
'patch' is not recognized as an internal or external command,
operable program or batch file.
Array
(
)
Returned: 1
Time: 4.47 secs; Memory: 13.25Mb

Patch-set: 10