Update patch set 3
Patch Set 3:

Yep. The Release Engineering team is working on it. Bear in mind it is a task of epic proportions. There's a lot of work involved in it. It requires re-building the stack from the ground up. Probably another quarter to really get it running. And then we still have to set up git caching in the existing infrastructure, switch to smaller one-executor 2CPU instances; and port all existing jobs.

Patch-set: 3
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0