Revert "Use Trusty for mediawiki-core-jslint"
This reverts commit e94262d6d4d5757d1dbbc75735b45dcc18ab0405.

Changing the label to Trusty wasn't forgotten. The binary in question
was compiled for Precise, not Trusty.

Considering this binary is deprecated, please keep it as-is. Migrate
to npm jobs instead. For check jobs, it can stay as-is without
issues. Migrating this has little to no benefit. If we do want to migrate
it has to be reinstalled in integration/jenkins.git on a clean Trusty
machine first.

Change-Id: Ia084515931aeb5acaf47d52e782b2516f8c46f66