Update patch set 2
Patch Set 2:

That has nothing to do with Trusty/Precise. That has to do with labs/production. More specifically, that relates to the 'defaults' property in how it fetches from Git.

On the old production servers, there was /srv/ssd/gerrit. On the new labs slaves, it should fetch from zuul-merger.

That's a separate issue not solved by moving to Trusty. Both before and after revert that will be broken.

Patch-set: 2