Update patch set 2
Patch Set 2:

No, the problem was not fixed when this moved to Trusty.

I assume you are referring to this build which showed that /srv/ssd/gerrit error: https://integration.wikimedia.org/ci/job/mediawiki-core-jslint/12951/console

The reason that job failed is not actually related to "/srv/ssd/gerrit". That job failed because zuul.eqiad.wmnet timed out after 10 seconds.

That was most likely related to unrelated maintenance or traffic spike and fixed the next build.

The "/srv/ssd/gerrit" error just means it tries that one first and then it tries labs. That's normal.

Patch-set: 2